+++ idli_book/idli_book/doctype/ib_country/ib_country.js
@@ -1,0 +1,10 @@
+// Copyright (c) 2026, Praveen and contributors
+// For license information, please see license.txt
+
+frappe.ui.form.on('IB Country', {
+    refresh: function (frm) {
+        if (frm.doc.__islocal) {
+            frm.set_intro(__('Create a new country'));
+        }
+    }
+});